move tests out of src, so they don't go in dist
